requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.3
psycopg[binary]==3.2.1
django-jsonform==2.23.0
python-dotenv==1.0.1
django-redis==6.0.0
redis==6.4.0
django-parler